refactor: extract logic for setting up test envs
E2E tests can be executed with an external shared running app (that must be
started before running all the tests) or with an isolated env per test
(that can be customized and it lasts only until the end of the test).
