+++ tests/e2e/contexts/settings/contract.rs
@@ -2,7 +2,7 @@ use crate::common::contexts::settings::form::UpdateSettingsForm;
 use crate::common::contexts::settings::responses::{AllSettingsResponse, Public, PublicSettingsResponse, SiteNameResponse};
 use crate::common::contexts::settings::{Auth, Database, Mail, Net, Settings, Tracker, Website};
 use crate::e2e::contexts::user::steps::logged_in_admin;
-use crate::e2e::environment::TestEnv;
+use crate::environments::shared::TestEnv;
 
 #[tokio::test]
 #[cfg_attr(not(feature = "e2e-tests"), ignore)]
