+++ tests/environments/mod.rs
@@ -1,0 +1,3 @@
+pub mod app_starter;
+pub mod isolated;
+pub mod shared;
