+++ tests/integration/mod.rs
@@ -1,3 +1,1 @@
-pub mod app_starter;
-mod contexts;
-pub mod environment;
+pub mod contexts;
