+++ tests/mod.rs
@@ -1,5 +1,6 @@
 mod common;
 mod databases;
 mod e2e;
+pub mod environments;
 mod integration;
 mod upgrades;
